LayoutTests/fast/canvas/webgl/incorrect-context-object-behaviour-expected.txt

Tests calling WebGL APIs with objects from other contexts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS contextA.compileShader(shaderB) threw exception GL error 1282 in compileShader.
PASS contextA.linkProgram(programB) threw exception GL error 1282 in linkProgram.
PASS contextA.attachShader(programA, shaderB) threw exception GL error 1281 in attachShader.
PASS contextA.attachShader(programB, shaderA) threw exception GL error 1281 in attachShader.
PASS contextA.attachShader(programB, shaderB) threw exception GL error 1281 in attachShader.
PASS contextA.detachShader(programA, shaderB) threw exception GL error 1282 in detachShader.
PASS contextA.detachShader(programB, shaderA) threw exception GL error 1282 in detachShader.
PASS contextA.detachShader(programB, shaderB) threw exception GL error 1282 in detachShader.
PASS contextA.shaderSource(shaderB, 'foo') threw exception GL error 1282 in shaderSource.
PASS contextA.bindAttribLocation(programB, 0, 'foo') threw exception GL error 1282 in bindAttribLocation.
PASS contextA.bindFramebuffer(contextA.FRAMEBUFFER, frameBufferB) threw exception GL error 1282 in bindFramebuffer.
PASS contextA.bindRenderbuffer(contextA.RENDERBUFFER, renderBufferB) threw exception GL error 1282 in bindRenderbuffer.
PASS contextA.bindTexture(contextA.TEXTURE_2D, textureB) threw exception GL error 1282 in bindTexture.
PASS contextA.framebufferRenderbuffer(contextA.FRAMEBUFFER, contextA.DEPTH_ATTACHMENT, contextA.RENDERBUFFER, renderBufferB) threw exception GL error 1282 in framebufferRenderbuffer.
PASS contextA.framebufferTexture2D(contextA.FRAMEBUFFER, contextA.COLOR_ATTACHMENT0, contextA.TEXTURE_2D, textureB, 0) threw exception GL error 1282 in framebufferTexture2D.
PASS contextA.getProgramParameter(programB, 0) threw exception GL error 1282 in getProgramParameter.
PASS contextA.getProgramInfoLog(programB, 0) threw exception GL error 1282 in getProgramInfoLog.
PASS contextA.getShaderParameter(shaderB, 0) threw exception GL error 1282 in getShaderParameter.
PASS contextA.getShaderInfoLog(shaderB, 0) threw exception GL error 1282 in getShaderInfoLog.
PASS contextA.getShaderSource(shaderB) threw exception GL error 1282 in getShaderSource.
PASS contextA.getUniform(programB, locationA) threw exception GL error 1282 in getUniform.
PASS contextA.getUniformLocation(programB, 'u_modelViewProjMatrix') threw exception GL error 1282 in getUniformLocation.
PASS successfullyParsed is true

TEST COMPLETE
